Updating tests to support tokens in `str` format
Details:

The previous commit `89e88e4` fixes shorter token generation and pads the
generated tokens with appropriate number of zeroes on the left. In order to
support this, generated tokens had to be produced as string. This commit updates
the `test.py` to run the token verification tests with strings as opposed to
integers.