Improve constant-time comparison by avoiding short-circuiting
This PR improves constant-time verification for the functions
`utils._compare_digest()` and `totp.verify()`.

In the old `util._compare_digest()`, once a comparison returns a non-zero value
the loop continues, but short-circuits before evaluating all the comparisons
that should follow. By writing `ord(c1) ^ ord(c2) | differences` instead of
`difference | ord(c1) ^ ord(c2)` (which is what `|=` is syntactic sugar for), we
actually evaluate the XOR expression each loop for equal-length strings.

In the old `totp.verify()`, when `valid_window` is specified, the verification
will short-circuit if for any window the OTP is valid. In the case that an
adversary enters a valid username and OTP for that user, but not a valid
password, it seems reasonable that we don't want to leak that the OTP was valid.