Have random_base32() use 'secrets' as rand source
In PEP 506 https://www.python.org/dev/peps/pep-0506/ it's recommended (suggested) that "(...) secrets becomes the go-to module for dealing with anything which should remain secret (passwords, tokens, etc.) (...)"

secrets was included in python 3.6. This update will use secrets providing the most secure way to generate the token if secrets is available, and will default to random if not.